+++ frontend/js/app.js
@@ -1,7 +1,6 @@
 import { getArticles, getArticleById, getReadingProgress } from './api.js';
 import { initHighlightFeature } from './highlight.js';
 import { closeOriginSnippetPanel, closeReader, openOriginSnippetPanel, renderReader } from './reader.js';
-import { initReferenceTestPanel } from './reference.js';
 
 const state = {
   tab: 'today',
@@ -255,7 +254,6 @@ function init() {
   }
 
   bindEvents();
-  initReferenceTestPanel(showToast);
   initHighlightFeature({
     readerContent: nodes.readerContent,
     getCurrentArticle: () => state.currentArticle,